% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{consult}
\alias{consult}
\title{Consult a prolog database}
\usage{
consult(fname = system.file("pl/likes.pl", package = "rolog"))
}
\arguments{
\item{fname}{file name of database}
}
\value{
\code{TRUE} on success
}
\description{
Consult a prolog database
}
\examples{
consult(fname=system.file('pl/likes.pl', package='rolog'))
findall(likes(sam, X))

}
\seealso{
\code{\link[=once]{once()}}, \code{\link[=findall]{findall()}}, and \code{\link[=query]{query()}}/\code{\link[=submit]{submit()}}/\code{\link[=clear]{clear()}} for executing queries
}
