% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{once}
\alias{once}
\title{Invoke a query once}
\usage{
once(
  query = call("member", expression(X), list(1, 2, 3)),
  portray = TRUE,
  tovec = FALSE
)
}
\arguments{
\item{query}{an R call}

\item{portray}{boolean, add the prolog translation as an attribute}

\item{tovec}{boolean. Vectors of characters, integers, and floats are
translated to compounds of $, \% and #, respectively. Vectors of
length 1 are translated to scalars unless tovec is set to TRUE.}
}
\value{
\code{FALSE} if the query fails; otherwise, a list with conditions
}
\description{
Invoke a query once
}
\examples{

# This query returns FALSE
once(call("=", 1, 2))

# This query returns an empty list meaning yes, it works
once(call("=", 1, 1))

# This query returns a list stating that it works if X = 1
once(call("member", 1, list(2, expression(X))))

# This query returns a list stating that X = 1 and Z = Y
once(call("=", list(expression(X), expression(Y)), list(1, expression(Z))))

# works for X = [1 | _]; i.e. something like [|](1, expression(_6330))
once(call("member", 1, expression(X)))

}
\seealso{
\code{\link[=findall]{findall()}} for querying all solutions

\code{\link[=portray]{portray()}} for pretty-printing a query
}
