% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{postproc}
\alias{postproc}
\title{Default hook for postprocessing}
\usage{
postproc(constraint = call("=<", 1, 2))
}
\arguments{
\item{constraint}{the R call representing constraints of the Prolog query.}
}
\value{
The default hook translates the inequality and smaller-than-or-equal-to back
from Prolog (\=, =<) to R (!=, <=).
}
\description{
Default hook for postprocessing
}
\seealso{
[rolog_options()] for fine-grained control over the translation
}
