% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{preproc}
\title{Default hook for preprocessing}
\usage{
preproc(query = quote(1 <= 2))
}
\arguments{
\item{query}{the R call representing the Prolog query.}
}
\value{
The default hook translates the inequality and smaller-than-or-equal-to from
R (!=, <=) to Prolog (\=, =<).
}
\description{
Default hook for preprocessing
}
\seealso{
[rolog_options()] for fine-grained control over the translation
}
