\name{rolog_consult}
\alias{rolog_consult}
\docType{package}
\title{
Consult Prolog file(s)
}
\description{
This function consults the Prolog program \code{prolog_files}.
}
\usage{
rolog_consult(prolog_files)	
}
\arguments{
  \item{prolog_files}{The prolog file(s) to be consulted.}
}
\examples{
\dontrun{
rolog_consult("buggy.pl")
}
}
