% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{rolog_done}
\alias{rolog_done}
\title{rolog_done}
\usage{
rolog_done()
}
\value{
\code{TRUE} on success
}
\description{
Clean up when detaching the library. At this stage, this function is of
little practical use, since it is not yet possible to initialize prolog
twice in the same R session. See the source file rolog.cpp for details.
}
