\name{rolog_init}
\alias{rolog_init}
\docType{package}
\title{
Initialize the prolog interpreter
}
\description{
This global function initializes the prolog interpreter. It should be called only once in a session.
}
\usage{
rolog_init(home, prc)	
}
\arguments{
  \item{home}{The prolog path.}
  \item{prc}{The prolog init file.}
}
\examples{
\dontrun{
rolog_init()
}
}
