% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{rolog_ok}
\alias{rolog_ok}
\title{Check if rolog is properly loaded}
\usage{
rolog_ok(warn = FALSE, stop = FALSE)
}
\arguments{
\item{warn}{raise a warning if problems occurred}

\item{stop}{raise an error if problems occurred}
}
\value{
TRUE if rolog is properly loaded
}
\description{
Check if rolog is properly loaded
}
