% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{rolog_options}
\alias{rolog_options}
\title{Quick access to rolog's own options}
\usage{
rolog_options()
}
\value{
list with some options for translating R expressions to Prolog
}
\description{
Quick access to rolog's own options
}
\details{
Translation of R to Prolog
\itemize{
\item numeric vector of size N -> $realvec/N (default is #)
\item integer vector of size N -> $intvec/N (default is \%)
\item boolean vector of size N -> $boolvec/n (default is !)
\item character vector of size N -> $charvec/N (default is $)
\item $scalar: if TRUE (default), translate R vectors of size 1 to scalars in
Prolog
\item $portray: whether to return the prolog translation as an attribute to
once and findall (default is TRUE)
}
}
