% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{rolog_quote}
\alias{rolog_quote}
\title{Translate simplified to canonical representation}
\usage{
rolog_quote(x = quote(member(X, list(a, "b", 3L, 4, TRUE, Y))))
}
\arguments{
\item{query}{an R call representing a prolog query with prolog-like syntax, e.g.,
`member(X, list(a, b, Y))` that is used in [query()], [once()], and [findall()].
This is query is translated to Rolog's canonical representation, with
R calls and prolog variables enclosed in an R expression, in this example,
`call("member", expression(X), list(a, b, expression(Y))))`.}
}
\description{
Translate simplified to canonical representation
}
\examples{
query(quote(member(X, list(a, "b", 3L, 4, TRUE, Y))))

}
\seealso{
[query()], [once()], [findall()]
which use this function if `options$quote` is `TRUE`.

[rolog_options()]
option _quote_ controls whether this function is used.
}
