% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usemodule.R
\name{usemodule}
\alias{usemodule}
\title{Use a prolog module}
\usage{
usemodule(fname = system.file(file.path("pl", "family.pl"), package = "rolog"))
}
\arguments{
\item{fname}{file name of module}
}
\value{
\code{TRUE} on success
}
\description{
Use a prolog module
}
\examples{
usemodule(fname=system.file(file.path("pl", "family.pl"), package="rolog"))
findall(call("ancestor", quote(pam), expression(X)))

}
\seealso{
\code{\link[=consult]{consult()}} for consulting a prolog file
}
