% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{consult}
\alias{consult}
\title{Consult a prolog database}
\usage{
consult(fname = system.file("likes.pl", package = "rolog"))
}
\arguments{
\item{fname}{file name of database}
}
\value{
\code{TRUE} on success
}
\description{
Consult a prolog database
}
\examples{
consult(fname=system.file("likes.pl", package="rolog"))
findall(call("likes", quote(sam), expression(X)))

}
\seealso{
\code{\link[=once]{once()}} and \code{\link[=findall]{findall()}} for executing queries
}
