% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{findall}
\alias{findall}
\title{Invoke a query several times}
\usage{
findall(query = call("member", expression(X), list(1, 2, 3)), options = NULL)
}
\arguments{
\item{query}{an R call}

\item{options}{boolean. See rolog_options}
}
\value{
empty list if the query fails; otherwise, a list of conditions for
each solution
}
\description{
Invoke a query several times
}
\examples{
findall(call("member", expression(X), list(1, 2, 3)))

}
\seealso{
\code{\link[=once]{once()}} for a single query

\code{\link[=rolog_options]{rolog_options()}}
}
