% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{findall}
\alias{findall}
\title{Invoke a query several times}
\usage{
findall(query = call("member", expression(X), list(1, 2, 3)), options = NULL)
}
\arguments{
\item{query}{an R call, consisting of symbols (= prolog atoms),
numbers (= prolog numbers), strings (= prolog strings),
boolean values (= prolog atoms true and false),
expressions (= prolog variables) and lists (= prolog lists), and other
calls (= prolog compounds). Vectors of booleans, integers, floating point
numbers, and strings with length \emph{N} > 1 are translated to prolog
compounds !/N, \%/N, #/N and $/N, respectively. The names can be modified
with the options below.}

\item{options}{list of options controlling translation from and to prolog:
boolvec (see option rolog.boolvec, default is !) is the name of the
prolog compound for boolean vectors. intvec, realvec and charvec define
the compound names for vectors of integers, doubles and strings,
respectively (defaults are \%, # and $). If \emph{scalar} is TRUE (default),
vectors of length 1 are translated to scalar prolog elements. If \emph{scalar}
is FALSE, even vectors of length 1 are translated to compounds.}
}
\value{
If the query fails, an empty list is returned. If the query
succeeds \emph{N} >= 1 times, a list of length \emph{N} is returned, each element
being a list of conditions for each solution.
}
\description{
Invoke a query several times
}
\examples{
findall(call("member", expression(X), list(1, 2, 3)))

}
\seealso{
\code{\link[=once]{once()}} for a single query

\code{\link[=rolog_options]{rolog_options()}}
}
