% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{once}
\alias{once}
\title{Invoke a query once}
\usage{
once(query = call("member", expression(X), list(1, 2, 3)), portray = TRUE)
}
\arguments{
\item{query:}{an R call}

\item{portray:}{boolean, add the prolog translation as an attribute}
}
\value{
\code{FALSE} if the query fails; otherwise, a list with conditions
}
\description{
Invoke a query once
}
\examples{
once(call("=", 1, 2)) # FALSE 
once(call("=", 1, 1)) # empty list
once(call("member", 1, list(2, expression(X)))) # list stating that it works if X = 1
once(call("=", list(expression(X), expression(Y)), list(1, expression(Z)))) # list stating that X = 1 and Z = Y
once(call("member", 1, expression(X))) # works for X = [1 | _]; i.e. something like [|](1, expression(_6330))

}
\seealso{
\code{\link[=findall]{findall()}} for querying all solutions

\code{\link[=portray]{portray()}} for pretty-printing a query
}
