% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{rolog_init}
\alias{rolog_init}
\title{Start prolog}
\usage{
rolog_init(argv1 = commandArgs()[1])
}
\arguments{
\item{argv1}{file name of the R executable}
}
\value{
`TRUE` on success
}
\description{
Start prolog
}
\details{
SWI-prolog is automatically initialized when the rolog library is loaded, so
this function is normally not directly invoked.
}
