% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolog.R
\name{rolog_options}
\alias{rolog_options}
\title{Quick access the package options}
\usage{
rolog_options()
}
\value{
list with some options for translating R expressions to prolog
}
\description{
Quick access the package options
}
\details{
Translation from R to Prolog
\itemize{
\item numeric vector of size N -> \emph{realvec}/N (default is ##)
\item integer vector of size N -> \emph{intvec}/N (default is \%\%)
\item boolean vector of size N -> \emph{boolvec}/N (default is !!)
\item character vector of size N -> \emph{charvec}/N (default is $$)
\item \emph{scalar}: if \code{TRUE} (default), translate R vectors of length 1 to scalars
\item \emph{portray}: if \code{TRUE} (default) whether to return the prolog translation
as an attribute to the return value of \code{\link[=once]{once()}}, \code{\link[=query]{query()}} and \code{\link[=findall]{findall()}}
}
}
